treewide: standardize environment variable format
Previously, I had named the environment variables based on whatever felt
right in the moment. However, this led to many of the names being unclear
to what the variable actually set. The messages also didn't clarify which
program the environment variable was actually for, so if you set them,
forgot about them, and saw them in your list of environment variables, you
might have been confused.

Now, all the environment variables start with the name of the program (
RBLD or UNIFY). I also reworded some of the names slightly to be clearer.