unify: always print logging when restoring `flake.lock`
Previously, we had a step to restore the `flake.lock` both in the trap,
and within the normal script execution. However, only one of these places
actually had logging. We now only have the operating for restoring in
one place: the trap. The trap checks if the `flake.lock` actually needs
to be restored, and if so, it prints what it's doing, and restores it.
Now, we'll get the same output whether the script finishes, or if we
`Ctrl+c` it.