unify: always cleanup state
Previously, I had our trap named as `return_to_secondary`. But now, we're
using it for cleaning up any state that generates during execution.
Sometimes, we're on main and never needed to swap, but we should still be
using the trap for getting rid of flake.lock state.

As well as renaming the function here, we now *always* clean up state.
The last commit ensured we didn't swap branches if we didn't need to, but
it had the unintended effect of making us not always cleanup state. Now,
we sign up for state cleanup outside of the if statement.

There may be a nanosecond of time between execution of
`switch_to_primary` and the `trap`, where state could hypothetically be
preserved. I'm choosing not to fix this, because if you're able to hit
Ctrl+C during this millisecond, I'm choosing to keep this, because I
think it's more impressive to be able to Ctrl+C during this time. If
you're able to do it, tell me, I'll be very impressed!