fuiska: fix buggy behavior with non-branch inputs
First, `ls-remote --branches` failed to work for inputs that pointed to
HEAD. To fix that, we just have an if statement that ensures we only use
`--branches` when necessary. We also exit early in case newHash is
empty, which happens when `ls-remote` fails for some reason. If I had
this before, I would've noticed this bug earlier!