fuiska/fight: add api option
If $MENU_USE_API is set to anything other than false, or fuiska is
called with `fuiska --api` or `fuiska -a`, it'll use the Github/Gitlab
api where possible. Note that Github is very stingy about rate limiting,
and I haven't implemented a way to supply your GH API key (and I
 wouldn't want to make that required by default). I also don't take
advantage of any API features yet - the only feature right now is it's
faster than the non-API fetch.