fight: handle both annotated and lightweight tags
My previous change to unpeel all tags with `^{}` worked for one example,
but it turns out that it broke "lightweight" tags, that don't need
unpeeling. I could've resolved this via a simple `if` statement, or some
more `-z "$newHash"`, but I was tired of nesting, so I found a cool way.

We first get the output of `git ls-remote`, but adding `*` to the end
of the ref. This will show only the tag and its rev if it's a
lightweight tag, but if it's a annotated tag, there'll be two lines -
one for the unpeeled, the second for the peeled. We then use `cut -f1`
to strip off the metadata.

Now, we either have one line with the rev, or two lines, with the rev we
want on the second line. Fish splits on newlines, so the annnotated tag
will become a list with two elements. But, fish treats ALL variables as
lists. So, the lightweight tag becomes a list with one element. Notice
how in both cases, the important data is at the last element? Well,
here's the kicker. Fish ALSO supports negative indices. So we just grab
the -1st index, and set that as our `newHash`. We did it! This accounts
for both lightweight and annotated tags.

They say the real programming happens when you run out of memory. But I
say the real programming comes when you start placing unnecessary
stylistic constraints on yourself.