fight: add tarball support
The whole thing that makes fuiska faster than Nix (other than
parallelism) is that it doesn't fetch tarballs - it just queries
metadata from a git repo with `git ls-remote`. But, how can we handle
flake inputs that just point to a tarball link? They're pointing to a
tarball, not a git repo, so we can't do the usual querying.

Well, for MOST tarball links you'll find in flake inputs, they're not
just pointing to some generic "https://web.site/file.tar.gz". They're
pointing to an archive or release from a git repo. This means that for
most links, we can try to reconstruct them into a git repo. Regex can
help us here! Links that don't match will still be ignored by fuiska,
but doing this will make fuiska work for *most* tarball links.

As this code will probably outlast my memory of it, here's what the
regex does, bit by bit:

`(https?:\/`
Begins a capturing group for the base URL, andMatches either the string
https:/", or "http:/". Note that we don't include the second slash yet.
This is for reasons you'll see later.

`(?:\/[^\/]+){3})`
We start with a non-capturing group, that grabs a slash, followed by a
tring up until another slash. Examples include "github.com/",
"aaaaaaaa/", etc. We end this non-capturing group, and match for it
three times. This would match "/github.com/llakala/menu", which is
matched thrice as "/github.com", "/llakala", and "/menu". This is why
we didn't include the second slash in "https://": we wanted to match
that slash here. We do this three times, to match a website name,
followed by the user, followed by the repo. Finally, we end the first
capturing group. This is the base URL that we'll use with git ls-remote.

`\/(?:archive|releases\/download)\/`
A slash, followed by one of two strings, followed by another slash. Only
two possibilities here: "/archive/", or "/releases/download/". The
archive logic is meant for most tarballs, with the /releases/download
logic only for dealing with tarballs that point to a github release.

`(?:refs\/tags\/)?`
Matches the string "refs/tags" with a non-capturing group, but let the
group be optional. More Github logic - sometimes, a tarball URL includes
this. For those URLs, this just skips us ahead to the right point in the
URL.

`([^\/]+)`
Our second and last capturing group. This grabs any non-slash character
any number of times. We store this as our ref to use with git
ls-remote. For a URL like this:
"https://github.com/verilator/verilator/archive/refs/tags/v3.916.tar.gz"",
this would match "v3.916" and use it as our ref.

`(?:\/[^\/]+)?`
Another optional string, handled by a non-capturing group and a `?`.
This is where we get into the weird kind of URLs. Refer to the verilator
example above. It would be almost done at this point, with only a
".tar.gz" left. But, for inputs matching the "releases/download" style
like this one:
"https://github.com/KaTeX/KaTeX/releases/download/v0.16.9/katex.zip",
we aren't done yet. To handle this, we grab a slash, followed by a
string of non-slashes. I don't actually understand why this needs to be
a string of non-slashes. But when I try it with any other set of
characters, it fails and matches too much. So, we'll take it. In our
example, this would match "/katex". You may ask, why doesn't it match
`.zip`? Keep reading to learn why!

`(?:\.tar\.gz|\.zip|\.tar\.xz)`
Very simple logic here. We either match `.tar.gz`, `.zip`, or `.tar.xz`.
Perhaps there's some formats that some URLs will use, that will be
ignored because of this. But I wanted to specify what I knew would work.
Feel free to PR in some more examples, or make it smarter.