fuiska/fight: only run regex once
We use an intermediary variable `both`, as a list of two elements - the
URL and the ref. Since we need to be in double quotes for the --replace,
we can't use something like `$1\n$2` (or, at least, I don't know of a
way to do that without also treating $1 and $2 literally). Instead, we
just use a space as a separator. Should be fine - if an input URL or ref
somehow had a space in it, someone's messed up there.